requirements.txt
# MongoDB driver
pymongo>=4.6.0

# OpenAI client (for Yandex AI)
openai>=1.0.0

# Telegram Bot API
python-telegram-bot>=20.0

# Progress bars
tqdm>=4.65.0

# Environment variable loading
python-dotenv>=1.0.0

# Playwright for web scraping
playwright>=1.40.0
scrapy-playwright>=0.0.36
